requirements.txt
python-telegram-bot[job-queue]==20.1
deep-translator==1.10.1
deepl==1.14.0
requests==2.28.2
psycopg2-binary==2.9.5
python-dotenv==1.0.0
orjson==3.8.6
CairoSVG==2.6.0
numpy==1.24.2